remove LogOut
